imu draft implementation

+++ main/include/types.hpp
@@ -25,7 +25,7 @@ typedef struct {
     uint16_t mag_x;
     uint16_t mag_y;
     uint16_t mag_z;
-    uint8_t temp;
+    uint16_t temp;
 } imu_reading_t;
 
 typedef uint32_t rtc_reading_t;
